requirements.txt
mem0ai>=0.1.76
langchain-openai>=0.0.5
langchain-core>=0.1.0
python-dotenv>=1.0.0
langgraph>=0.0.15 